==> 1_hyperparametres fasttext.ipynb <==
This notebook is dedicated to the fine tuning of the fasttext model.

We use here a previous work done by one of the authors : https://github.com/alexandredupuy-zini/dynamic-topic-modeling. Make sure to modify the parameters.yml file located at dynamic-topic-modeling/conf/base with the one given in this repository, and to modify them in order to fit your subject.

This additional repository is not maintained, and therefore needs to be used with caution as it requires an old version of kedro. The kedro pipeline was originally built with Kedro 0.15.4.

import os

os.chdir('dynamic-topic-modeling/')

os.environ['NUMEXPR_MAX_THREADS'] = '45'

!python -m kedro run --pipeline grid_search_fasttext

# fasttext grid_search exploration

import pandas as pd
import pickle
import matplotlib.pyplot as plt
import numpy as np

di = pickle.load(open("data/04_features/grid_search_results.pkl", "rb"))

pd.DataFrame.from_dict(di)

## ws=7 sounds better

fig, ax = plt.subplots(nrows=1, ncols=3, sharey=True, figsize=(20,5))

ax[0].plot(df.groupby('ws').mean()['ndgc'])
ax[1].plot(df.groupby('epoch').mean()['ndgc'])
ax[2].plot(df.groupby('minCount').mean()['ndgc'])

ax[0].set_xlabel('window size')
ax[1].set_xlabel('epochs')
ax[2].set_xlabel('minimum count')

plt.show()

plt.plot(df.groupby('ws').mean()['ndgc'])
plt.show()

## 10 ou 15 epoch are best

plt.plot(df.groupby('epoch').mean()['ndgc'])
plt.show()

## minCount has little influence, except fixing the vocabulary size

plt.plot(df.groupby('minCount').mean()['ndgc'])
plt.show()

# Variance

minCount doesn't influence much the performances, the higher ws is the better for the `hit` ratio, and epochs positvely impact `ndgc`.

Vocabulary is more relevant with a higher ws, and using more epoch makes the links between relevant words stronger.

for col in ['ws', 'epoch', 'minCount']:
    print(col)
    for x in ['hit', 'ndgc']:
        print("\t{}\t : {}".format(x, np.var(df.groupby(col).mean()[x])))

# Best hyperparameters selection

df[df['ws']==7][df['hit']>=0.958333]['ndgc'].argmax()

df.iloc[16]

df.iloc[df['hit'].argmax()]

## Final Hyperparameters choice

df.iloc[df['ndgc'].argmax()]

df.iloc[df['hit'].argmax()]['cos_sim']
